For this task, I used logic to see the issue and corrected it. So not really code related maybe it was some luck.
Since I saw that it was producing a 16x16 square, and and saw in the "size" variable" there was a 16, I changed it by 10 without really knowing what I was changing.

And it worked out !

But after digging I understand, this was to specify if it was hexadecimal, decimal or else bases. So I stated then the base was decimal and so 10 = 10 and not 16