In this test, there were two issues, both related to the password: - The first was in the password setter. It was just a tiny spelling mistake, self.__password was set to "None" at the begining of the code as a global variable and in the setter the password given was meant to be hashed and stored in it. BUT instead of "self.__password" it was stored in "self._password" which isn't the same attribute. Leading to not hashing the password at all.
- Second issue was in the test "valid_password". To sum up the comparison was set in "uppercase" before comparing, while in the hasher it had been set in lowercaser (both with .lower and .upper) so they had to be the same to do the same comparision.

NOTE: While juste changing .upper to .lower to be the same as in the hasher. I actually removed these two at all. Because in my opinion changing by default the password into lower or uppercase is a security weakness because it means that anyway the user (or hacker) test in upper or lowercase it will be in lowercase in the end. Removing a layer of security. So I removed both lower and upper to match EXACTLY the password